requirements.txt

pillow
datetime
pyjokes
pyautogui
pywhatkit
wolframalpha
PyQt5
SpeechRecognition
geopy
geocoder
pyttsx3
secure-smtplib
psutil
google-generativeai